requirements.txt
Django>=1.3
PIL==1.1.7
south>=0.7.3
BeautifulSoup==3.2.0
django_compressor==0.9.2
django-debug-toolbar>=0.8.5
django-templatetag-sugar>=0.1
django-taggit>=0.9.3
